Mutualize code for forms to send emails. Remove thousand line of code.

+++ htdocs/core/actions_massactions.inc.php
@@ -64,11 +64,9 @@ if (! $error && $massaction == 'confirm_presend')
     $langs->load("mails");
     include_once DOL_DOCUMENT_ROOT.'/core/lib/files.lib.php';
 
-    if (!$error && !isset($user->email))
-    {
-        $error++;
-        setEventMessages($langs->trans("NoSenderEmailDefined"), null, 'warnings');
-    }
+    $listofobjectid=array();
+    $listofobjectthirdparties=array();
+    $listofobjectref=array();
 
     if (! $error)
     {
@@ -76,12 +74,9 @@ if (! $error && $massaction == 'confirm_presend')
         if ($objecttmp->element == 'expensereport') $thirdparty=new User($db);
 
         $objecttmp=new $objectclass($db);
-        $listofobjectid=array();
-        $listofobjectthirdparties=array();
-        $listofobjectref=array();
         foreach($toselect as $toselectid)
         {
-            $objecttmp=new $objectclass($db);	// must create new instance because instance is saved into $listofobjectref array for future use
+            $objecttmp=new $objectclass($db);	// we must create new instance because instance is saved into $listofobjectref array for future use
             $result=$objecttmp->fetch($toselectid);
             if ($result > 0)
             {
@@ -93,10 +88,39 @@ if (! $error && $massaction == 'confirm_presend')
                 $listofobjectref[$thirdpartyid][$toselectid]=$objecttmp;
             }
         }
-        //var_dump($listofobjectthirdparties);exit;
+    }
 
+    // Check mandatory parameters
+    if (empty($user->email))
+    {
+        $error++;
+        setEventMessages($langs->trans("NoSenderEmailDefined"), null, 'warnings');
+        $massaction='presend';
+    }
 
-        // Loop on each thirdparty
+    $receiver=$_POST['receiver'];
+    if (! is_array($receiver))
+    {
+    	if (empty($receiver) || $receiver == '-1') $receiver=array();
+    	else $receiver=array($receiver);
+    }
+    if (count($receiver) == 0 && count($listofobjectthirdparties) == 1)	// if only one recipient, receiver is mandatory
+    {
+     	$error++;
+       	setEventMessages($langs->trans("ErrorFieldRequired", $langs->transnoentitiesnoconv("Recipient")), null, 'warnings');
+       	$massaction='presend';
+    }
+
+    if (! GETPOST('subject','none'))
+    {
+    	$error++;
+    	setEventMessages($langs->trans("ErrorFieldRequired", $langs->transnoentitiesnoconv("MailTopic")), null, 'warnings');
+    	$massaction='presend';
+    }
+
+    // Loop on each recipient/thirdparty
+    if (! $error)
+    {
         foreach ($listofobjectthirdparties as $thirdpartyid)
         {
             $result = $thirdparty->fetch($thirdpartyid);
@@ -112,12 +136,6 @@ if (! $error && $massaction == 'confirm_presend')
             $sendtoid = array();
 
             // Define $sendto
-            $receiver=$_POST['receiver'];
-            if (! is_array($receiver))
-            {
-            	if ($receiver == '-1') $receiver=array();
-            	else $receiver=array($receiver);
-            }
             $tmparray=array();
             if (trim($_POST['sendto']))
             {
@@ -433,10 +451,10 @@ if (! $error && $massaction == 'confirm_presend')
             //setEventMessages($langs->trans("EMailSentToNRecipients", 0), null, 'warnings');  // May be object has no generated PDF file
             setEventMessages($resaction, null, 'warnings');
         }
-    }
 
-    $action='list';
-    $massaction='';
+        $action='list';
+    	$massaction='';
+    }
 }
 
 if (! $error && $massaction == "builddoc" && $permtoread && ! GETPOST('button_search'))
@@ -618,7 +636,7 @@ if ($action == 'remove_file')
     $action='';
 }
 
-// Validate  records
+// Validate records
 if (! $error && $massaction == 'validate' && $permtocreate)
 {
 	if ($object->element == 'invoice_supplier' && ! empty($conf->stock->enabled) && ! empty($conf->global->STOCK_CALCULATE_ON_SUPPLIER_BILL))
@@ -690,8 +708,19 @@ if (! $error && $massaction == 'delete' && $permtodelete)
     	$result=$objecttmp->fetch($toselectid);
         if ($result > 0)
         {
+        	// Refuse deletion for some status ?
+        	/*
+       		if ($objectclass == 'Facture' && $objecttmp->status == Facture::STATUS_DRAFT)
+       		{
+       			$langs->load("errors");
+       			$nbignored++;
+       			$resaction.='<div class="error">'.$langs->trans('ErrorOnlyDraftStatusCanBeDeletedInMassAction',$object->ref).'</div><br>';
+       			continue;
+       		}*/
+
             if (in_array($objecttmp->element, array('societe','member'))) $result = $objecttmp->delete($objecttmp->id, $user, 1);
             else $result = $objecttmp->delete($user);
+
             if ($result <= 0)
             {
                 setEventMessages($objecttmp->error, $objecttmp->errors, 'errors');
